+++ mbsebbs/mbtoberep.c
@@ -2,7 +2,7 @@
  *
  * File ..................: mbtoberep/mbtoberep.c
  * Purpose ...............: Show contents of toberep.data
- * Last modification date : 28-Jun-2001
+ * Last modification date : 09-Aug-2001
  *
  *****************************************************************************
  * Copyright (C) 1997-2001
@@ -69,7 +69,7 @@ int main(int argc, char **argv)
 		printf("Comment       %s\n", rep.Comment);
 		printf("Group         %s\n", rep.Group);
 		printf("File name     %s\n", rep.Name);
-		printf("FIle size     %lu\n", rep.Size);
+		printf("FIle size     %lu\n", (long)(rep.Size));
 		printf("File size Kb  %lu\n", rep.SizeKb);
 		printf("File date     %s", ctime(&rep.Fdate));
 		printf("File CRC      %s\n", rep.Crc);
